% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python.R
\docType{data}
\name{boto}
\alias{boto}
\title{boto module}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
boto
}
\description{
boto module
}
\keyword{datasets}
