% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boto_helpers.R
\name{ec2_instance_create}
\alias{ec2_instance_create}
\title{ec2_instance_create}
\usage{
ec2_instance_create(image_id = NA, instance_type = "t2.nano",
  min = 1, max = 1, key_name = NA, security_group_id = NA,
  instance_storage = 50, device_name = "/dev/sda1", user_data = NA)
}
\arguments{
\item{image_id}{An aws ec2 id: i.e., 'ami-0174e69c12bae5410'}

\item{instance_type}{See \url{https://aws.amazon.com/ec2/instance-types/}}

\item{min}{min instances}

\item{max}{max instances}

\item{key_name}{A .pem file to ssh}

\item{security_group_id}{SecurityGroupId of security group you have created in UI}

\item{instance_storage}{Size of the box in gb}

\item{device_name}{"/dev/sda1"}

\item{user_data}{A shell script that runs on startup}
}
\description{
ec2_instance_create
}
