% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boto_helpers.R
\name{security_group_create}
\alias{security_group_create}
\title{security_group_create}
\usage{
security_group_create(group_key = NA)
}
\arguments{
\item{group_name}{Name of security group}
}
\description{
security_group_create
}
